Apply refactor to gdeep
