+++ gdeep/trainer/pipeline_tool/class_impl/LayerClass.py
@@ -1,5 +1,5 @@
 import torch
-from ..contant import TAB
+from ..constant import TAB
 
 class Layer:
     """Set up of the Layer to manage every thing for creating a layer with the init and forward functions.
